refactor: Move batch tx logic into hook
