+++ src/components/transactions/BatchExecuteButton/index.tsx
@@ -1,55 +1,17 @@
-import { useCallback, useContext, useMemo } from 'react'
+import { useCallback, useContext } from 'react'
 import { Button } from '@mui/material'
 import css from './styles.module.css'
 import { BatchExecuteHoverContext } from '@/components/transactions/BatchExecuteButton/BatchExecuteHoverProvider'
 import { Transaction, TransactionListItem } from '@gnosis.pm/safe-react-gateway-sdk'
-import useSafeInfo from '@/hooks/useSafeInfo'
-import { isMultisigExecutionInfo, isTransactionListItem } from '@/utils/transaction-guards'
 import { useAppSelector } from '@/store'
 import { selectPendingTxs } from '@/store/pendingTxsSlice'
 import CustomTooltip from '@/components/common/CustomTooltip'
-
-const BATCH_LIMIT = 10
-
-export const getBatchableTransactions = (items: (TransactionListItem | Transaction[])[], nonce: number) => {
-  const batchableTransactions: Transaction[] = []
-  let currentNonce = nonce
-
-  const grouped = items
-    .map((item) => {
-      if (Array.isArray(item)) return item
-      if (isTransactionListItem(item)) return [item]
-    })
-    .filter(Boolean) as Transaction[][]
-
-  grouped.forEach((txs) => {
-    const sorted = txs.slice().sort((a, b) => b.transaction.timestamp - a.transaction.timestamp)
-    sorted.forEach((tx) => {
-      if (!isMultisigExecutionInfo(tx.transaction.executionInfo)) return
-
-      const { nonce, confirmationsSubmitted, confirmationsRequired } = tx.transaction.executionInfo
-      if (
-        batchableTransactions.length < BATCH_LIMIT &&
-        nonce === currentNonce &&
-        confirmationsSubmitted >= confirmationsRequired
-      ) {
-        batchableTransactions.push(tx)
-        currentNonce = nonce + 1
-      }
-    })
-  })
-
-  return batchableTransactions
-}
+import useBatchedTxs from '@/hooks/useBatchedTxs'
 
 const BatchExecuteButton = ({ items }: { items: (TransactionListItem | Transaction[])[] }) => {
   const pendingTxs = useAppSelector(selectPendingTxs)
   const hoverContext = useContext(BatchExecuteHoverContext)
-  const { safe } = useSafeInfo()
-
-  const currentNonce = safe.nonce
-
-  const batchableTransactions = useMemo(() => getBatchableTransactions(items, currentNonce), [currentNonce, items])
+  const batchableTransactions = useBatchedTxs(items)
 
   const isBatchable = batchableTransactions.length > 1
   const hasPendingTx = batchableTransactions.some((tx) => pendingTxs[tx.transaction.id])
